More Javadoc and code style fixes
