+++ src/main/java/org/tudo/sse/resolution/PomResolutionException.java
@@ -6,10 +6,16 @@ import org.tudo.sse.model.ArtifactIdent;
  * This class handles any issues that arise during pom resolution, giving more informative error messages.
  */
 public class PomResolutionException extends Exception {
+
     private final String message;
     private final ArtifactIdent artifactIdentifier;
 
-
+    /**
+     * Creates a new PomResolutionException with the given attributes detailing the exception.
+     * @param message A message describing the error
+     * @param artifactIdentifier An artifact identifier referencing the affected artifact
+     * @param cause The underlying cause, if available - else null.
+     */
     public PomResolutionException(String message, ArtifactIdent artifactIdentifier, Throwable cause) {
         super(cause);
         this.message = message;
