+++ src/main/java/org/tudo/sse/resolution/ResolverFactory.java
@@ -12,21 +12,39 @@ import org.tudo.sse.model.ArtifactIdent;
  * This class manages the pom and jar resolver, giving a way to run one or the other.
  */
 public class ResolverFactory {
+
     private final PomResolver pomResolver;
     private final JarResolver jarResolver;
 
-    public static final Logger log = LogManager.getLogger(ResolverFactory.class);
+    private static final Logger log = LogManager.getLogger(ResolverFactory.class);
 
+    /**
+     * Creates a new resolver factory instance. Resolvers will not output the artifacts that they process.
+     *
+     * @param pomIncludeTransitives Whether this factory shall produce a PomResolver that resolves transitive files.
+     */
     public ResolverFactory(boolean pomIncludeTransitives) {
         pomResolver = new PomResolver(pomIncludeTransitives);
         jarResolver = new JarResolver();
     }
 
+    /**
+     * Creates a new resolver factory instance.
+     *
+     * @param output Whether the resolvers produced by this factory shall output the artifacts that they process
+     * @param pathToDirectory Location to which to output the artifacts processed by the resolvers
+     * @param pomIncludeTransitives Whether this factory shall produce a PomResolver that resolves transitive files
+     */
     public ResolverFactory(boolean output, Path pathToDirectory, boolean pomIncludeTransitives) {
         pomResolver = new PomResolver(output, pathToDirectory, pomIncludeTransitives);
         jarResolver = new JarResolver(output, pathToDirectory);
     }
 
+    /**
+     * Resolve the POM file of the given artifact.
+     *
+     * @param identifier Artifact identifier to resolve
+     */
     public void runPom(ArtifactIdent identifier) {
         try {
             pomResolver.resolveArtifact(identifier);
@@ -35,6 +53,11 @@ public class ResolverFactory {
         } catch (FileNotFoundException ignored) {}
     }
 
+    /**
+     * Resolve the JAR file of the given artifact.
+     *
+     * @param identifier Artifact identifier to resolve
+     */
     public void runJar(ArtifactIdent identifier) {
         try {
             jarResolver.parseJar(identifier);
@@ -43,6 +66,11 @@ public class ResolverFactory {
         }
     }
 
+    /**
+     * Resolve both the POM and JAR file for the given artifact.
+     *
+     * @param identifier Artifact identifier to resolve
+     */
     public void runBoth(ArtifactIdent identifier) {
         try {
             pomResolver.resolveArtifact(identifier);
